combined geospatial.py with validate_inputs_spatial.py
